requirements.txt
asgiref==3.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-froala-editor==4.0.9
django-tinymce==3.4.0
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2021.5
